sudah berhasil melakukan pembelian lewat cart

+++ app/Services/Transaksi/OrderService.php
@@ -6,6 +6,7 @@ use App\Models\Pembeli;
 use App\Models\Produk\Stok;
 use Illuminate\Support\Collection;
 use Illuminate\Support\Facades\DB;
+use App\Models\Produk\ProdukVariant;
 use Illuminate\Support\Facades\Http;
 
 class OrderService
@@ -40,23 +41,36 @@ class OrderService
 		}
 	}
 
-	private function initMidtrans(array $data, $orderId):array
+	private function initMidtrans(array|Collection $data,$username,$totalHarga, $orderId):array
 	{
-		$pembeli = Pembeli::where('username', '=',$data['username'])->first();
+		$pembeli = Pembeli::where('username', '=',$username)->first();
+		$items = [];
+		if(count($data) == 1){
+			foreach($data as $value){
+				$items[] = [
+					'price' => $value['hargaSatuan'],
+					'quantity' => $value['jumlah'],
+					'name' => $orderId
+				];
+			}
+		}else{
+			foreach($data as $row){
+				$variant = ProdukVariant::where('id', '=', $row->variant_id)->first();
+				$items[] = [
+					'price' => $variant->harga,
+					'quantity' => $row->qty,
+					'name' => $orderId
+ 				];
+			}
+		}
 		$params = [
             'transaction_details' => [
                 'order_id' => $orderId,
-                'gross_amount' => $data['totalHarga']
+                'gross_amount' => $totalHarga
             ],
-            'item_details' => [
-                [
-                    'price' => $data['hargaSatuan'],
-                    'quantity' => $data['jumlah'],
-                    'name' => $orderId
-                ],
-            ],
+            'item_details' => $items,
             'customer_details'=> [
-                'first_name' => $data['username'],
+                'first_name' => $username,
                 'email' => $pembeli->email
             ],
             'enable_payments' => ['credit_card', 'bni_va', 'bca_va', 'gopay', 'alfamart', 'indomart'],
@@ -113,8 +127,13 @@ class OrderService
 	        $lastInsertOrderItemsId = DB::getPdo()->lastInsertId();
 			// stok baru di update, saat user menyelesaikan pembayaran
 	        // DB::statement("UPDATE stoks set jumlah = jumlah - ? WHERE variant_id = ?", [$data['jumlah'], $data['variantId']]);
-
-            $response = $this->initMidtrans($data, $orderId);
+			$items = [
+				[
+					'hargaSatuan' => $data['hargaSatuan'],
+					'jumlah' => $data['jumlah']
+				]	
+			];
+            $response = $this->initMidtrans($items,$data['username'], $data['totalHarga'] ,$orderId);
             if (!isset($response['redirect_url'])) {
 			    DB::rollback();
 			    $error = $response['status_message'] ?? 'Gagal membuat transaksi Midtrans';
@@ -146,6 +165,7 @@ class OrderService
 		$isSuccess = false;
 		$userId = $data['userId'];
 		$totalHarga = $data['totalHarga'];
+		$username = $data['username'];
 
 		DB::beginTransaction();
 
@@ -165,12 +185,23 @@ class OrderService
         	$tanggalSekarang = now()->format('Y-m-d');
 
         	$this->insertIntoOrderTable($tanggalSekarang, $data['userId'], $data['totalHarga'], $orderInsertId, $orderId);
-
 			DB::table('table_orders')->where('id', $orderInsertId)->update([
 				'order_id' => $orderId
 			]);
-
+			
 			$this->insertIntoOrderItems($carts, $orderInsertId);
+			$response = $this->initMidtrans($carts, $username, $totalHarga, $orderId);
+
+			if (!isset($response['redirect_url'])) {
+			    DB::rollback();
+			    $error = $response['status_message'] ?? 'Gagal membuat transaksi Midtrans';
+			    return false;
+			}
+            $linkBayar = $response['redirect_url'];
+            DB::table('table_orders')->where('id', '=', $orderInsertId)->update([
+            	'order_id' => $orderId,
+            	'link_bayar' => $response['redirect_url'] ?? null
+            ]);
 
 			DB::table('carts')->where('pembeli_id', $userId)->delete();
 			DB::commit();
